+++ geo-types/src/line_string.rs
@@ -63,7 +63,7 @@ use std::ops::{Index, IndexMut};
 /// ```
 /// use geo_types::LineString;
 ///
-/// let line_string: LineString<f64> = vec![[0., 0.], [10., 0.]].into();
+/// let line_string: LineString = vec![[0., 0.], [10., 0.]].into();
 /// ```
 //
 /// Or by `collect`ing from a [`Coordinate`] iterator
@@ -132,7 +132,7 @@ use std::ops::{Index, IndexMut};
 
 #[derive(Eq, PartialEq, Clone, Debug, Hash)]
 #[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
-pub struct LineString<T: CoordNum>(pub Vec<Coordinate<T>>);
+pub struct LineString<T: CoordNum = f64>(pub Vec<Coordinate<T>>);
 
 /// A [`Point`] iterator returned by the `points` method
 #[derive(Debug)]