+++ geo/benches/geodesic_distance.rs
@@ -6,8 +6,8 @@ use geo::prelude::*;
 
 fn criterion_benchmark(c: &mut criterion::Criterion) {
     c.bench_function("geodesic distance f64", |bencher| {
-        let a = geo::Point::<f64>::new(17.107558, 48.148636);
-        let b = geo::Point::<f64>::new(16.372477, 48.208810);
+        let a = geo::Point::new(17.107558, 48.148636);
+        let b = geo::Point::new(16.372477, 48.208810);
 
         bencher.iter(|| {
             criterion::black_box(