+++ geo/src/algorithm/relate/geomgraph/intersection_matrix.rs
@@ -226,7 +226,7 @@ impl IntersectionMatrix {
     /// use geo_types::{LineString, Rect, line_string};
     /// use geo::{coordinate_position::CoordPos, dimensions::Dimensions, relate::Relate};
     ///
-    /// let line_string: LineString<f64> = line_string![(x: 0.0, y: 0.0), (x: 10.0, y: 0.0), (x: 5.0, y: 5.0)];
+    /// let line_string: LineString = line_string![(x: 0.0, y: 0.0), (x: 10.0, y: 0.0), (x: 5.0, y: 5.0)];
     /// let rect = Rect::new((0.0, 0.0), (5.0, 5.0));
     ///
     /// let intersection = line_string.relate(&rect);