+++ geo/CHANGES.md
@@ -4,6 +4,12 @@
 
 * Reexport everything from the `proj` crate
   * <https://github.com/georust/geo/pull/839>
+* Add densification algorithm for linear geometry components
+  * <https://github.com/georust/geo/pull/847>
+* Fix fast path euclidean distance
+  * <https://github.com/georust/pull/848>
+* Use robust predicates everywhere in geo
+  * <https://github.com/georust/geo/pull/852>
 
 ## 0.21.0